Add files via upload
$1.50 each
Fresh Bunch of Bananas,
Avg. 2.48 lbs

Description: 
Pick up a perfect option for a quick breakfast or on-the-go snacking with this Banana. A good source of potassium, the banana is great for adding to oatmeal, cereal, ice cream and more. Plus, you can also blend it with yogurt and other favorite fruits to whip up a delicious smoothie.