Add files via upload
Fresh Mango,
Price : $0.50 each
Available Qty : 100 nos
Brand : Mango Royale
Description:
Mangoes vary in size, shape, and color, but they are typically oval or round with a smooth skin that ranges from green to various shades of yellow, orange, or red when ripe. The skin may have small blemishes or spots. The taste is a harmonious balance of sweetness and tartness, creating a unique and refreshing flavor profile